Improve exp loading
