+++ packages/expo-cli/src/exp.ts
@@ -1,6 +1,5 @@
 import bunyan from '@expo/bunyan';
 import { setCustomConfigPath } from '@expo/config';
-import { INTERNAL_CALLSITES_REGEX } from '@expo/metro-config';
 import boxen from 'boxen';
 import chalk from 'chalk';
 import program, { Command } from 'commander';
@@ -18,8 +17,6 @@ import {
   ApiV2,
   Binaries,
   Config,
-  ConnectionStatus,
-  Doctor,
   Logger,
   LogRecord,
   LogUpdater,
@@ -344,6 +341,7 @@ Command.prototype.asyncAction = function (asyncFn: Action) {
     try {
       const options = args[args.length - 1];
       if (options.offline) {
+        const { ConnectionStatus } = await import('xdl');
         ConnectionStatus.setIsOffline(true);
       }
 
@@ -369,6 +367,7 @@ Command.prototype.asyncAction = function (asyncFn: Action) {
         } else {
           Log.addNewLineIfNone();
           Log.error(err.message);
+          const { formatStackTrace } = await import('./utils/formatStackTrace');
           const stacktrace = formatStackTrace(err.stack, this.name());
           Log.error(chalk.gray(stacktrace));
         }
@@ -381,63 +380,6 @@ Command.prototype.asyncAction = function (asyncFn: Action) {
     }
   });
 };
-
-function getStringBetweenParens(value: string): string {
-  const regExp = /\(([^)]+)\)/;
-  const matches = regExp.exec(value);
-  if (matches && matches?.length > 1) {
-    return matches[1];
-  }
-  return value;
-}
-
-function focusLastPathComponent(value: string): string {
-  const parts = value.split('/');
-  if (parts.length > 1) {
-    const last = parts.pop();
-    const current = chalk.dim(parts.join('/') + '/');
-    return `${current}${last}`;
-  }
-  return chalk.dim(value);
-}
-
-function formatStackTrace(stacktrace: string, command: string): string {
-  const treeStackLines: string[][] = [];
-  for (const line of stacktrace.split('\n')) {
-    const [first, ...parts] = line.trim().split(' ');
-    // Remove at -- we'll use a branch instead.
-    if (first === 'at') {
-      treeStackLines.push(parts);
-    }
-  }
-
-  return treeStackLines
-    .map((parts, index) => {
-      let first = parts.shift();
-      let last = parts.pop();
-
-      // Replace anonymous with command name
-      if (first === 'Command.<anonymous>') {
-        first = chalk.bold(`expo ${command}`);
-      } else if (first?.startsWith('Object.')) {
-        // Remove extra JS types from function names
-        first = first.split('Object.').pop()!;
-      } else if (first?.startsWith('Function.')) {
-        // Remove extra JS types from function names
-        first = first.split('Function.').pop()!;
-      } else if (first?.startsWith('/')) {
-        // If the first element is a path
-        first = focusLastPathComponent(getStringBetweenParens(first));
-      }
-
-      if (last) {
-        last = focusLastPathComponent(getStringBetweenParens(last));
-      }
-      const branch = (index === treeStackLines.length - 1 ? '└' : '├') + '─';
-      return ['   ', branch, first, ...parts, last].filter(Boolean).join(' ');
-    })
-    .join('\n');
-}
 
 // asyncActionProjectDir captures the projectDirectory from the command line,
 // setting it to cwd if it is not provided.
@@ -518,7 +460,9 @@ Command.prototype.asyncActionProjectDir = function (
       }
     };
 
-    const logStackTrace = (
+    const { INTERNAL_CALLSITES_REGEX } = await import('@expo/metro-config');
+
+    const logStackTrace = async (
       chunk: LogRecord,
       logFn: (...args: any[]) => void,
       nestedLogFn: (...args: any[]) => void
@@ -693,6 +637,7 @@ Command.prototype.asyncActionProjectDir = function (
       Log.setSpinner(spinner);
       // validate that this is a good projectDir before we try anything else
 
+      const { Doctor } = await import('xdl');
       const status = await Doctor.validateWithoutNetworkAsync(projectRoot, {
         skipSDKVersionRequirement: options.skipSDKVersionRequirement,
       });
